# api/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: BizCon API
  version: 0.1.0
  description: |
    API for BizCon Admin and User Chat.
    - Public endpoints are accessible to anyone
    - User chat endpoints are accessible by UUID
    - Admin endpoints are restricted to localhost
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key for admin endpoints. Note that admin endpoints are also restricted to localhost access.
  schemas:
    Lead:
      type: object
      properties:
        id: { type: string, description: "Unique lead/user ID (UUID)" }
        name: { type: string }
        company: { type: string }
        email: { type: string }
        description: { type: string }
        source: { type: string }
        email_sent: { type: boolean }
        email_sent_count: { type: integer }
        last_email_sent: { type: string, format: date-time }
        private_link: { type: string, description: "User's private chat link" }
        status: { type: string, enum: [Hot, Warm, Cold, Not Responded] }
        chat_summary: { type: string }
        connected: { type: boolean }
...
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        error: 
          type: string
          description: Error message describing what went wrong
...
paths:
  /api/upload/company-files:
    post:
      summary: Upload company PDF files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        file:
                          type: string
                        status:
                          type: string
                          enum: [success, error]
                        error:
                          type: string
                          nullable: true
        '400':
          description: Bad request - No files provided
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/upload/company-urls:
    post:
      summary: Upload company URLs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                urls:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        url:
                          type: string
                        status:
                          type: string
                          enum: [success, error]
                        error:
                          type: string
                          nullable: true
        '400':
          description: Bad request - No URLs provided
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/upload/user-files:
    post:
      summary: Upload user (leads) files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        file:
                          type: string
                        status:
                          type: string
                          enum: [success, error]
                        leads:
                          type: integer
                          nullable: true
                        error:
                          type: string
                          nullable: true
        '400':
          description: Bad request - No files provided
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/settings:
    get:
      summary: Get all settings
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Whether the operation was successful
                  settings:
                    type: object
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/settings/email:
    post:
      summary: Save email settings
      description: |
        Save and validate email settings. The credentials will be validated before saving.
        - Tests SMTP connection to verify credentials
        - Returns validation errors if credentials are invalid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender
                - password
              properties:
                sender:
                  type: string
                  description: Email sender address
                password:
                  type: string
                  description: Email password or app-specific password
                host:
                  type: string
                  description: SMTP host (optional)
                port:
                  type: integer
                  description: SMTP port (optional)
      responses:
        '200':
          description: Settings saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Bad request - Invalid or missing data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/settings/azure:
    post:
      summary: Save Azure settings
      description: |
        Save and validate Azure OpenAI settings. The credentials will be validated before saving.
        - Tests both embeddings and chat completion connections
        - Returns validation errors if credentials are invalid or services are unreachable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - endpoint
                - api_key
                - api_version
                - deployment
                - embedding_deployment
              properties:
                endpoint:
                  type: string
                  description: Azure OpenAI endpoint
                api_key:
                  type: string
                  description: Azure OpenAI API key
                api_version:
                  type: string
                  description: Azure OpenAI API version
                deployment:
                  type: string
                  description: Azure OpenAI deployment name
                embedding_deployment:
                  type: string
                  description: Azure OpenAI embedding deployment name
      responses:
        '200':
          description: Settings saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Bad request - Invalid or missing data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
  /api/leads:
    get:
      summary: Get grouped leads
      description: Returns all leads grouped by source.
      responses:
        '200':
          description: Grouped leads
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/Lead'
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/clear-all:
    post:
      summary: Clear all data (chats, files, and leads)
      description: Clears all stored data including chat histories, uploaded files, and lead information
      responses:
        '200':
          description: Clear result
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Failed to clear data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...